restfull webservice - spring boot/mobile-app-ws/src/test/java/app/ws/shared/UtilsTest.java
package app.ws.shared;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import static org.junit.Assert.*;

@ExtendWith(SpringExtension.class)
@SpringBootTest
class UtilsTest {

    @Autowired
    Utils utils;

    @BeforeEach
    void setUp() {

    }

    @Test
    final void generateUserId() {
        String userId = utils.generateUserId(30);
        String anotherUserId = utils.generateUserId(30);
        assertNotNull(userId);
        assertTrue(userId.length() == 30);
        assertTrue(userId.equalsIgnoreCase(anotherUserId));
    }

    @Test
    final void hasTokenNotExpired() {
        String token = utils.generateEmailVerificationToken("4futrd4A655f1gsfg");
        assertNotNull(token);
        boolean hasTokenExpired = Utils.hasTokenExpired(token);
        assertFalse(hasTokenExpired);
    }

    @Test
    final void hasTokenExpired() {
        String token = "";
        boolean hasTokenExpired = Utils.hasTokenExpired(token);
        assertTrue(hasTokenExpired);
    }
}